Revert "single device"
This reverts commit 813838ddb6618216f66b4f08ed829f763b02cea5.

+++ python/sgl_jax/srt/utils/weight_utils.py
@@ -1068,10 +1068,9 @@ class WeightLoader:
 
                 return _load_slice
 
-            lazy_array = jax.make_array_from_callback(
-                shape, sharding, _make_load_slice(), dtype=target_dtype
-            ).astype(target_dtype)
-
+            lazy_array = jax.make_array_from_callback(shape, sharding, _make_load_slice()).astype(
+                target_dtype
+            )
 
             lazy_arrays.append(lazy_array)
